I only wrote tests for my github methods. I hope that is ok. Can always add the other ones later
