+++ packages/graphql_server/src/test/githubApi.test.ts
@@ -1,4 +1,4 @@
-import { getRepositoryTopics } from '../scraping/githubScraping'
+import { getRepositoryTopics } from '../api/githubApi'
 
 function testGetRepositoryTopics() {
   //get real comments from method in hackerNewsScraper. E.g. searchHackerNewsStories('SuperAGI')
