+++ packages/graphql_server/types/hackerNewsScraping.d.ts
